+++ src/rez/bind/hello_world.py
@@ -10,7 +10,7 @@ from __future__ import absolute_import, print_function
 from rez.package_maker__ import make_package
 from rez.vendor.version.version import Version
 from rez.utils.lint_helper import env
-from rez.util import create_executable_script, ExecutableScriptMode
+from rez.utils.execution import create_executable_script, ExecutableScriptMode
 from rez.bind._utils import make_dirs, check_version
 import os.path
