+++ src/rez/bind/rezgui.py
@@ -8,7 +8,7 @@ from rez.bind._utils import check_version, make_dirs
 from rez.system import system
 from rez.vendor.version.version import Version
 from rez.utils.lint_helper import env
-from rez.util import create_executable_script
+from rez.utils.execution import create_executable_script
 import shutil
 import os.path
